MHD-Kontrolle — desktop E2E: shelf-life checkbox on the picking-profile window (replaces stale C_Workplace test)
